+++ src/force/nep3.cuh
@@ -18,9 +18,9 @@
 #include "utilities/gpu_vector.cuh"
 
 struct NEP3_Data {
-  GPU_Vector<double> f12x; // 3-body or manybody partial forces
-  GPU_Vector<double> f12y; // 3-body or manybody partial forces
-  GPU_Vector<double> f12z; // 3-body or manybody partial forces
+  GPU_Vector<float> f12x; // 3-body or manybody partial forces
+  GPU_Vector<float> f12y; // 3-body or manybody partial forces
+  GPU_Vector<float> f12z; // 3-body or manybody partial forces
   GPU_Vector<float> Fp;
   GPU_Vector<float> sum_fxyz;
   GPU_Vector<int> NN_radial;    // radial neighbor list
