+++ src/force/potential.cuh
@@ -45,9 +45,9 @@ protected:
     Box& box,
     const int* NN,
     const int* NL,
-    const double* f12x,
-    const double* f12y,
-    const double* f12z,
+    const float* f12x,
+    const float* f12y,
+    const float* f12z,
     const GPU_Vector<double>& position_per_atom,
     GPU_Vector<double>& force_per_atom,
     GPU_Vector<double>& virial_per_atom);
